+++ src/code4/stdlib.h
@@ -4,6 +4,6 @@
 #ifndef STDLIB_H
 #define STDLIB_H
 
-void itoa(unsigned long int n, unsigned short int* buffer, int basenumber);
+void itoa(unsigned long int n, unsigned long int* buffer, unsigned long int basenumber);
 
 #endif // STDLIB_H
